Stack.c
This code represents a basic implementation of a stack data structure in C using dynamic memory allocation, which prevents the stack from having a potential end. Of course, the limit will depend on the operating system it will be executed on and its hardware. A stack is a linear data structure that follows the Last In, First Out (LIFO) principle, meaning that the last element added to the stack is the first one to be removed.